Move the 'prompt_for_confirmation' function from 'users' to 'core' module.
